Logout handling (#69)
* Rename BBB API query key enum.
Introduce logout param.

* Better error handling in case of missing or invalid JWT.
Introduce logout handling. #44

+++ src/main/kotlin/de/thm/mni/ii/classroom/properties/JwtProperties.kt
@@ -18,7 +18,7 @@ data class JwtProperties(
     /**
      * validity duration of JWT in seconds
      */
-    @NotBlank @Range(min = 60) val expiration: Int,
+    @NotBlank @Range(min = 60) val expiration: Long,
 
     val jwtSubject: String = "classroom-auth"
